fix: rename and replace util-fn with similar one
- rename getContainerFromPoint to getCellContainerFromPoint
- replace getElementFromPoint with function getContainerFromPoint
