+++ src/behaviors/DefaultBehavior.ts
@@ -1,8 +1,9 @@
 import { Behavior } from "../types/Behavior";
-import { IndexedLocation } from "../types/InternalModel";
 import { handleKeyDown } from "../utils/handleKeyDown";
 import { hasTouchSupport, isMobile } from "../utils/isMobile";
 import { ReactGridStore } from "../utils/reactGridStore";
+import { getCellContainerLocation } from "../utils/getCellContainerLocation";
+import { getCellContainerFromPoint } from "../utils/getCellContainerFromPoint";
 
 type DefaultBehaviorConfig = {
   moveHorizontallyOnEnter: boolean;
@@ -17,35 +18,6 @@ let pointerDownPosition: { x: number; y: number } | null = null;
 let touchStartPosition: { x: number; y: number } | null = null;
 let touchEndPosition: { x: number; y: number } | null = null;
 
-const getElementFromPoint = (x: number, y: number): HTMLElement | null => {
-  const element = document.elementFromPoint(x, y) as HTMLElement;
-  if (element && element.classList.contains("rgCellContainer")) {
-    return element;
-  } else if (element?.closest(".rgCellContainer")) {
-    return element?.closest(".rgCellContainer");
-  }
-  return null;
-};
-
-function getCellContainerLocation(element: HTMLElement): IndexedLocation {
-  const rowIdxRegex = /rgRowIdx-(\d+)/;
-  const colIdxRegex = /rgColIdx-(\d+)/;
-
-  const rowIdxMatch = rowIdxRegex.exec(element.classList.value);
-  const colIdxMatch = colIdxRegex.exec(element.classList.value);
-
-  if (rowIdxMatch && colIdxMatch) {
-    return {
-      rowIndex: parseInt(rowIdxMatch[1]),
-      colIndex: parseInt(colIdxMatch[1]),
-    };
-  } else {
-    return {
-      rowIndex: -1,
-      colIndex: -1,
-    };
-  }
-}
 export const DefaultBehavior = (config: DefaultBehaviorConfig = CONFIG_DEFAULTS): Behavior => ({
   handlePointerDown: function (event, store): ReactGridStore {
     if (isMobile()) {
@@ -54,7 +26,7 @@ export const DefaultBehavior = (config: DefaultBehaviorConfig = CONFIG_DEFAULTS)
     console.log("DB/handlePointerDown");
 
     pointerDownPosition = { x: event.clientX, y: event.clientY };
-    const element = getElementFromPoint(event.clientX, event.clientY);
+    const element = getCellContainerFromPoint(event.clientX, event.clientY);
     let newRowIndex = -1;
     let newColIndex = -1;
     if (element) {
@@ -133,7 +105,7 @@ export const DefaultBehavior = (config: DefaultBehaviorConfig = CONFIG_DEFAULTS)
 
     // Disable moving (horizontal & vertical scrolling) if touchStartPosition is the same as focusedCell position.
     const focusedCell = store.getFocusedCell();
-    const element = getElementFromPoint(touchStartPosition.x, touchStartPosition.y);
+    const element = getCellContainerFromPoint(touchStartPosition.x, touchStartPosition.y);
     if (focusedCell && element) {
       const { rowIndex: touchRowIndex, colIndex: touchColIndex } = getCellContainerLocation(element);
 
@@ -178,8 +150,8 @@ export const DefaultBehavior = (config: DefaultBehaviorConfig = CONFIG_DEFAULTS)
     touchEndPosition = { x: clientX, y: clientY };
 
     if (touchStartPosition && touchEndPosition) {
-      const prevElement = getElementFromPoint(touchStartPosition.x, touchStartPosition.y);
-      const currElement = getElementFromPoint(touchEndPosition.x, touchEndPosition.y);
+      const prevElement = getCellContainerFromPoint(touchStartPosition.x, touchStartPosition.y);
+      const currElement = getCellContainerFromPoint(touchEndPosition.x, touchEndPosition.y);
       if (prevElement && currElement) {
         const prevCell = getCellContainerLocation(prevElement);
         const currCell = getCellContainerLocation(currElement);
